sui: split client and ptb by domain
